fix dtype of image data

+++ godot_rl/core/godot_env.py
@@ -358,7 +358,7 @@ class GodotEnv:
         shape,
     ):
         return (
-            np.frombuffer(bytes.fromhex(hex_string), dtype=np.float16)
+            np.frombuffer(bytes.fromhex(hex_string), dtype=np.uint8)
             .reshape(shape)
             .astype(np.float32)[:, :, :]  # TODO remove the alpha channel
         )
